hide bottom navigation when player is in create character mode and show tombstone icon next to character's name in notes
